requirements.txt
# ──────────────────────────────────────────────
# QA Automation Framework — Dependencies
# ──────────────────────────────────────────────

# Core testing
pytest>=7.4.0
pytest-xdist>=3.5.0

# Browser automation
playwright>=1.49.0
pytest-playwright>=0.5.0

# Reporting
allure-pytest>=2.13.0
pytest-html>=4.1.0

# API testing
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0

# Test data generation
faker>=20.0.0

# Utilities
python-dateutil>=2.9.0
